# rowexec: add encoded datums to HLL sketch