--- a/pkg/sql/randgen/encdatum.go
+++ b/pkg/sql/randgen/encdatum.go
@@ -12,6 +12,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/rowenc"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
+	"github.com/cockroachdb/errors"
 )
 
 // RandDatumEncoding returns a random DatumEncoding value.
@@ -88,17 +89,35 @@ func NullEncDatum() rowenc.EncDatum {
 	return rowenc.EncDatum{Datum: tree.DNull}
 }
 
+// DatumEncoding_NONE is only for tests. It is not a valid encoding. It is used
+// to instruct random datum generators to skip encoding the returned datum(s).
+const DatumEncoding_NONE catenumpb.DatumEncoding = -1
+
+func shouldEncode(enc catenumpb.DatumEncoding) bool {
+	return enc != DatumEncoding_NONE
+}
+
 // GenEncDatumRowsInt converts rows of ints to rows of EncDatum DInts.
 // If an int is negative, the corresponding value is NULL.
-func GenEncDatumRowsInt(inputRows [][]int) rowenc.EncDatumRows {
+func GenEncDatumRowsInt(inputRows [][]int, enc catenumpb.DatumEncoding) rowenc.EncDatumRows {
+	var a tree.DatumAlloc
 	rows := make(rowenc.EncDatumRows, len(inputRows))
 	for i, inputRow := range inputRows {
 		for _, x := range inputRow {
+			var d rowenc.EncDatum
 			if x < 0 {
-				rows[i] = append(rows[i], NullEncDatum())
+				d = NullEncDatum()
 			} else {
-				rows[i] = append(rows[i], IntEncDatum(x))
+				d = IntEncDatum(x)
 			}
+			if shouldEncode(enc) {
+				b, err := d.Encode(types.Int, &a, enc, nil /* appendTo */)
+				if err != nil {
+					panic(err)
+				}
+				d = rowenc.EncDatumFromEncoded(enc, b)
+			}
+			rows[i] = append(rows[i], d)
 		}
 	}
 	return rows
@@ -159,17 +178,80 @@ func MakeRepeatedIntRows(n int, numRows int, numCols int) rowenc.EncDatumRows {
 	return rows
 }
 
-// GenEncDatumRowsString converts rows of strings to rows of EncDatum Strings.
-// If a string is empty, the corresponding value is NULL.
-func GenEncDatumRowsString(inputRows [][]string) rowenc.EncDatumRows {
+// GenEncDatumRowsFloat converts rows of strings to rows of EncDatum DFloats.
+func GenEncDatumRowsFloat(inputRows [][]string, enc catenumpb.DatumEncoding) rowenc.EncDatumRows {
+	var a tree.DatumAlloc
 	rows := make(rowenc.EncDatumRows, len(inputRows))
 	for i, inputRow := range inputRows {
 		for _, x := range inputRow {
-			if x == "" {
-				rows[i] = append(rows[i], NullEncDatum())
-			} else {
-				rows[i] = append(rows[i], stringEncDatum(x))
+			f, err := tree.ParseDFloat(x)
+			if err != nil {
+				panic(errors.AssertionFailedf("could not parse float: %v", x))
 			}
+			d := rowenc.EncDatum{Datum: f}
+			if shouldEncode(enc) {
+				b, err := d.Encode(types.Float, &a, enc, nil /* appendTo */)
+				if err != nil {
+					panic(err)
+				}
+				d = rowenc.EncDatumFromEncoded(enc, b)
+			}
+			rows[i] = append(rows[i], d)
+		}
+	}
+	return rows
+}
+
+// GenEncDatumRowsString converts rows of strings to rows of EncDatum Strings.
+// If a string is empty, the corresponding value is NULL.
+func GenEncDatumRowsString(inputRows [][]string, enc catenumpb.DatumEncoding) rowenc.EncDatumRows {
+	var a tree.DatumAlloc
+	rows := make(rowenc.EncDatumRows, len(inputRows))
+	for i, inputRow := range inputRows {
+		for _, x := range inputRow {
+			var d rowenc.EncDatum
+			if x == "" {
+				d = NullEncDatum()
+			} else {
+				d = stringEncDatum(x)
+			}
+			if shouldEncode(enc) {
+				b, err := d.Encode(types.String, &a, enc, nil /* appendTo */)
+				if err != nil {
+					panic(err)
+				}
+				d = rowenc.EncDatumFromEncoded(enc, b)
+			}
+			rows[i] = append(rows[i], d)
+		}
+	}
+	return rows
+}
+
+// GenEncDatumRowsCollatedString converts rows of strings to rows of EncDatum
+// collated Strings.
+// If a string is empty, the corresponding value is NULL.
+func GenEncDatumRowsCollatedString(
+	inputRows [][]string, typ *types.T, enc catenumpb.DatumEncoding,
+) rowenc.EncDatumRows {
+	var a tree.DatumAlloc
+	rows := make(rowenc.EncDatumRows, len(inputRows))
+	for i, inputRow := range inputRows {
+		for _, x := range inputRow {
+			var d rowenc.EncDatum
+			if x == "" {
+				d = NullEncDatum()
+			} else {
+				d = collatedStringEncDatum(x, typ.Locale())
+			}
+			if shouldEncode(enc) {
+				b, err := d.Encode(typ, &a, enc, nil /* appendTo */)
+				if err != nil {
+					panic(err)
+				}
+				d = rowenc.EncDatumFromEncoded(enc, b)
+			}
+			rows[i] = append(rows[i], d)
 		}
 	}
 	return rows
@@ -180,13 +262,30 @@ func stringEncDatum(s string) rowenc.EncDatum {
 	return rowenc.EncDatum{Datum: tree.NewDString(s)}
 }
 
+// collatedStringEncDatum returns an EncDatum representation of a string.
+func collatedStringEncDatum(s string, locale string) rowenc.EncDatum {
+	d, err := tree.NewDCollatedString(s, locale, &tree.CollationEnvironment{})
+	if err != nil {
+		panic(err)
+	}
+	return rowenc.EncDatum{Datum: d}
+}
+
+func bytesEncDatum(b []byte) rowenc.EncDatum {
+	return rowenc.EncDatum{Datum: tree.NewDBytes(tree.DBytes(b))}
+}
+
 // GenEncDatumRowsBytes converts rows of bytes to rows of EncDatum value encoded
 // bytes.
-func GenEncDatumRowsBytes(inputRows [][][]byte) rowenc.EncDatumRows {
+func GenEncDatumRowsBytes(inputRows [][][]byte, enc catenumpb.DatumEncoding) rowenc.EncDatumRows {
 	rows := make(rowenc.EncDatumRows, len(inputRows))
 	for i, inputRow := range inputRows {
 		for _, x := range inputRow {
-			rows[i] = append(rows[i], rowenc.EncDatumFromEncoded(catenumpb.DatumEncoding_VALUE, x))
+			d := bytesEncDatum(x)
+			if shouldEncode(enc) {
+				d = rowenc.EncDatumFromEncoded(catenumpb.DatumEncoding_VALUE, x)
+			}
+			rows[i] = append(rows[i], d)
 		}
 	}
 	return rows