--- a/pkg/sql/rowexec/sampler.go
+++ b/pkg/sql/rowexec/sampler.go
@@ -23,6 +23,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/sql/sqltelemetry"
 	"github.com/cockroachdb/cockroach/pkg/sql/stats"
 	"github.com/cockroachdb/cockroach/pkg/sql/types"
+	"github.com/cockroachdb/cockroach/pkg/util/collatedstring"
 	"github.com/cockroachdb/cockroach/pkg/util/intsets"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/mon"
@@ -35,11 +36,12 @@ import (
 type sketchInfo struct {
 	spec execinfrapb.SketchSpec
 	// Exactly one of sketchOld and sketchNew will be set.
-	sketchOld *hllOld.Sketch
-	sketchNew *hllNew.Sketch
-	numNulls  int64
-	numRows   int64
-	size      int64
+	sketchOld            *hllOld.Sketch
+	sketchNew            *hllNew.Sketch
+	numNulls             int64
+	numRows              int64
+	size                 int64
+	legacyFingerprinting bool
 }
 
 // A sampler processor returns a random sample of rows, as well as "global"
@@ -100,6 +102,7 @@ func newSamplerProcessor(
 	post *execinfrapb.PostProcessSpec,
 ) (*samplerProcessor, error) {
 	useNewHLL := execversion.FromContext(ctx) >= execversion.V25_1
+	legacyFingerprinting := execversion.FromContext(ctx) < execversion.V25_2
 
 	// Limit the memory use by creating a child monitor with a hard limit.
 	// The processor will disable histogram collection if this limit is not
@@ -127,9 +130,10 @@ func newSamplerProcessor(
 	var sampleCols intsets.Fast
 	for i := range spec.Sketches {
 		s.sketches[i] = sketchInfo{
-			spec:     spec.Sketches[i],
-			numNulls: 0,
-			numRows:  0,
+			spec:                 spec.Sketches[i],
+			numNulls:             0,
+			numRows:              0,
+			legacyFingerprinting: legacyFingerprinting,
 		}
 		if useNewHLL {
 			s.sketches[i].sketchNew = hllNew.New14()
@@ -509,6 +513,115 @@ func (s *samplerProcessor) DoesNotUseTxn() bool {
 
 // addRow adds a row to the sketch and updates row counts.
 func (s *sketchInfo) addRow(
+	ctx context.Context, row rowenc.EncDatumRow, typs []*types.T, buf *[]byte,
+) (err error) {
+	if s.legacyFingerprinting {
+		return s.addRowLegacy(ctx, row, typs, buf)
+	}
+
+	s.numRows++
+	isNull := true
+	*buf = (*buf)[:0]
+	for _, col := range s.spec.Columns {
+		if b := row[col].EncodedBytes(); b != nil && !containsCollatedString(typs[col]) {
+			// If the datum is already encoded and does not contain a collated
+			// string, we can insert the encoded bytes directly into the sketch.
+			// Even though the encoded bytes contain the column ID and type,
+			// this will not break the "loose invariant" that equal datums will
+			// have equal bytes added to the sketch because we are always adding
+			// datums from the same table column, i.e., the column ID and type
+			// should be the same for every value in the column.
+			//
+			// Composite, value encoded datums may have different encodings for
+			// semantically equivalent types, so using the encoded bytes can
+			// skew the cardinality estimate slightly. This should be rare and
+			// hyperloglog cardinality is already an estimate, so it is
+			// considered acceptable. For floats, the only values affected are 0
+			// and -0, which are semantically equivalent but have different
+			// value encodings. For decimals, 0 and -0 are affected, as well as
+			// any equal values with different numbers of trailing zeros. JSON
+			// and array types containing decimals are also affected similarly.
+			//
+			// Value-encoded collated strings are more likely than other types
+			// to cause cardinality over-estimations because the ratio of
+			// physically distinct strings to semantically distinct strings can
+			// be much higher. For example, the und-u-ks-level2 locale is
+			// case-insensitive, so there are 8 different physical strings all
+			// equivalent to "foo": "foo", "Foo", "fOo", "foO", "FOo", "FoO",
+			// "fOO", and "FOO". For this reason, we fall-back to using
+			// Fingerprint for collated strings.
+			//
+			// TODO(mgartner): We should probably truncate b to some max size to
+			// prevent a really wide value from growing buf. Since the distinct
+			// count is an estimate anyway, truncating the value shouldn't have
+			// any real impact.
+			*buf = append(*buf, b...)
+		} else {
+			// Fallback to using the Fingerprint method to generate bytes to
+			// insert into the sketch.
+			//
+			// We pass nil DatumAlloc so that each datum allocation was
+			// independent (to prevent bounded memory leaks like we've seen in
+			// #136394). The problem in that issue was that the same backing
+			// slice of datums was shared across rows, so if a single row was
+			// kept as a sample, it could keep many garbage alive. To go around
+			// that we simply disabled the batching.
+			//
+			// We choose to not perform the memory accounting for possibly
+			// decoded tree.Datum because we will lose the references to row
+			// very soon.
+			*buf, err = row[col].Fingerprint(ctx, typs[col], nil /* da */, *buf, nil /* acc */)
+			if err != nil {
+				return err
+			}
+		}
+		if isNull {
+			// Avoid calling IsNull() if the datum is unset because it will
+			// panic. Instead return an assertion error that might help in
+			// debugging.
+			if row[col].IsUnset() {
+				return errors.AssertionFailedf("unset datum: col=%d row=%s", col, row.String(typs))
+			}
+			isNull = row[col].IsNull()
+		}
+		s.size += int64(row[col].DiskSize())
+	}
+
+	if isNull {
+		s.numNulls++
+	}
+	if s.sketchNew != nil {
+		s.sketchNew.Insert(*buf)
+	} else {
+		s.sketchOld.Insert(*buf)
+	}
+	return nil
+}
+
+// containsCollatedString returns true if the type is a collated string type
+// or a container type included a collated string type. It does not return
+// true with collated string types with a default-equivalent collation.
+func containsCollatedString(t *types.T) bool {
+	switch t.Family() {
+	case types.CollatedStringFamily:
+		return !collatedstring.IsDefaultEquivalentCollation(t.Locale())
+	case types.ArrayFamily:
+		return containsCollatedString(t.ArrayContents())
+	case types.TupleFamily:
+		for _, t := range t.TupleContents() {
+			if containsCollatedString(t) {
+				return true
+			}
+		}
+	}
+	return false
+}
+
+// addRowLegacy adds a row to the sketch and updates row counts. This is the
+// legacy implementation from versions prior to 25.2.
+//
+// TODO(mgartner): Remove this once compatibility with 25.1 is no longer needed.
+func (s *sketchInfo) addRowLegacy(
 	ctx context.Context, row rowenc.EncDatumRow, typs []*types.T, buf *[]byte,
 ) error {
 	var err error