#### sql: add BenchmarkAnalyze

Release note: None

#### sql/execversion: add V25_2

Release note: None

#### rowexec: add encoded datums to HLL sketch

Previously, the sampler would decode and re-encode all datums except for
integers with EncDatum.Fingerprint before adding them to the hyperloglog
sketch. This ensured that equal bytes were added to the sketch for two
semantically equivalent values, even if their encodings were different.
The overhead of decoding and re-encoding is significant, however.

Now, for all types except for collated strings or container types with
collated strings, the encoded bytes of the datum, if present, are added
directly to the hyperloglog sketch, skipping the overhead of
fingerprinting. See the comments added to `addRow` for more details.

```
name     old time/op    new time/op    delta
Analyze     216ms ± 3%     158ms ± 3%  -27.14%  (p=0.000 n=9+10)

name     old alloc/op   new alloc/op   delta
Analyze    75.6MB ± 4%    63.7MB ± 3%  -15.66%  (p=0.000 n=9+9)

name     old allocs/op  new allocs/op  delta
Analyze      573k ± 5%      355k ± 1%  -37.94%  (p=0.000 n=10+8)
```

Fixes #137362
Fixes #140652

Release note: None