@mgartner 2025-02-06
I measured a slight speed up in `BenchmarkSysbench` with table stats enabled, though hard to say if it's just noise. 

```
name                                  old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_read_write-8    4.96ms ± 0%    4.93ms ± 0%  -0.52%  (p=0.000 n=10+10)

name                                  old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_read_write-8    2.11MB ± 0%    2.12MB ± 0%  +0.48%  (p=0.000 n=10+9)

name                                  old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_read_write-8     9.90k ± 0%     9.91k ± 0%    ~     (p=0.225 n=10+10)
```