@mgartner 2025-02-12
*[`pkg/sql/logictest/testdata/logic_test/distsql_stats` line 3074 at r7](https://reviewable.io/reviews/cockroachdb/cockroach/140601#-OIuXZ7u6KhLycoRZERh:-OIvJ4fWA81v5a9W_xeP:b-rv3tvm) ([raw file](https://github.com/cockroachdb/cockroach/blob/c471adc5c9a93441ff56b23131e7088ee57dd1e9/pkg/sql/logictest/testdata/logic_test/distsql_stats#L3074)):*
<details><summary><i>Previously, mgartner (Marcus Gartner) wrote…</i></summary><blockquote>

This change is worrisome. I need to figure out why it's happening before merging.
</blockquote></details>

Becuase this changes distinct counts slightly, the outer buckets were not being created any more, causing the test changes. I updated the tests with injected stats to ensure there are outer buckets and make the test less likely to flake.


<!-- Sent from Reviewable.io -->